fix: Match member type with return type from library
SparkFun_AS3935::lightningEnergy() returns uint32_t

+++ src/SensorEvent.h
@@ -5,7 +5,7 @@ struct SensorEvent
 {
   uint8_t type;
   uint8_t distance;
-  long energy;
+  uint32_t energy;
 };
 
 #endif
